Fix an issue with Detox Expo initialization
Credit here goes to: [Travis Vignon](https://github.com/lookitsatravis)
for the fix.

The issue lies with the fact that the overriding of ENV variables to
specify the path to the detox hooks is done _after_ the Jest workers
have already been spun up, resulting in Detox waiting for the ready
signal that is never sent.

The full fix context is here:

https://github.com/expo/detox-tools/issues/1#issuecomment-740675824